refactor(reborn): localize script and mcp runtime adapters
